Add support for method level executorch backend config (#3266)
Summary:

There are use cases where we might like to supply a separate ExecutorchBackendConfig for each method in the model. An example use case is where we might want to alloc inputs for one method and not alloc them for another. In order to support this, in this diff we add support for passing in a dictionary of configs to `to_executorch`.

Reviewed By: cccclai

Differential Revision: D56499598